+++ popup/popup.js
@@ -7,4 +7,18 @@ document.addEventListener("DOMContentLoaded", function () {
         chrome.runtime.openOptionsPage();
         window.close();
     });
+
+    let donateCircleButton = document.getElementById('donateCircleButton');
+    donateCircleButton.setAttribute('title', chrome.i18n.getMessage("buyMeCoffee") ?? 'Support development');
+    donateCircleButton.addEventListener('click', function () {
+        window.close();
+        window.open('https://emvaized.diaka.ua/donate', '_blank');
+    });
+
+    let emailCircleButton = document.getElementById('emailCircleButton');
+    emailCircleButton.setAttribute('title', 'Email');
+    emailCircleButton.addEventListener('click', function () {
+        window.close();
+        window.open('mailto:maximtsyba@gmail.com');
+    });
 });